move update_supernode_pooling_config that is used only in the tutorial from core to tutorial.
